Progress on the language feature 4.0

+++ src/main/java/com/napier/sem/Language.java
@@ -15,7 +15,7 @@ public class Language
     /**
      * Country Population
      */
-    public Integer Population;
+    public long Population;
 
     /**
      * Language
